⚰️ 불필요한 코드를 제거한다
